Implement generic chart details component and use it in app/charts and fleet -> appco/charts
